fix: Alter table issue (#23)

+++ db2i_safeString.h
@@ -65,7 +65,7 @@ public:
 
 private: 
   char* buf;
-  uint64 curPos;
+  size_t curPos;
   size_t allocSize;
 };
